move functions of reddit_helpper to helper_funcs
