+++ gamestonk_terminal/behavioural_analysis/reddit_view.py
@@ -8,7 +8,7 @@ import praw
 import finviz
 from gamestonk_terminal.helper_funcs import check_positive, parse_known_args_and_warn
 from gamestonk_terminal import config_terminal as cfg
-from gamestonk_terminal.reddit_helpers import (
+from gamestonk_terminal.helper_funcs import (
     print_and_record_reddit_post,
     find_tickers,
 )
