+++ gamestonk_terminal/due_diligence/reddit_view.py
@@ -8,7 +8,7 @@ from psaw import PushshiftAPI
 import praw
 from gamestonk_terminal.helper_funcs import check_positive, parse_known_args_and_warn
 from gamestonk_terminal import config_terminal as cfg
-from gamestonk_terminal.reddit_helpers import print_and_record_reddit_post
+from gamestonk_terminal.helper_funcs import print_and_record_reddit_post
 
 
 def due_diligence(other_args: List[str], ticker: str):
